+++ Tests/Highstock/CreditsTest.php
@@ -7,7 +7,7 @@ use PHPUnit\Framework\TestCase;
 
 class CreditsTest extends TestCase
 {
-    protected $chart;
+    protected Highstock $chart;
     protected $credits;
 
     protected function setUp(): void
@@ -16,18 +16,18 @@ class CreditsTest extends TestCase
         $this->credits = $this->chart->credits;
     }
 
-    public function testEnabled()
+    public function testEnabled(): void
     {
         $this->credits->enabled(true);
         $this->assertTrue($this->credits->enabled);
-        $this->assertRegExp('/"enabled":true/', $this->chart->render());
+        $this->assertMatchesRegularExpression('/"enabled":true/', $this->chart->render());
 
         $this->credits->enabled(false);
         $this->assertFalse($this->credits->enabled);
-        $this->assertRegExp('/"enabled":false/', $this->chart->render());
+        $this->assertMatchesRegularExpression('/"enabled":false/', $this->chart->render());
     }
 
-    public function testHref()
+    public function testHref(): void
     {
         $link = "http://www.highcharts.com";
         $this->credits->href($link);
@@ -35,7 +35,7 @@ class CreditsTest extends TestCase
 //        $this->assertRegExp('/"href":"http:\/\/www\.highcharts\.com"/', $this->chart->render());
     }
 
-    public function testPosition()
+    public function testPosition(): void
     {
         $position = array(
             'align' => 'right',
@@ -46,10 +46,10 @@ class CreditsTest extends TestCase
 
         $this->credits->position($position);
         $this->assertEquals($this->credits->position, $position);
-        $this->assertRegExp('/"position":{"align":"right","x":-10,"verticalAlign":"bottom","y":-5}/', $this->chart->render());
+        $this->assertMatchesRegularExpression('/"position":{"align":"right","x":-10,"verticalAlign":"bottom","y":-5}/', $this->chart->render());
     }
 
-    public function testStyle()
+    public function testStyle(): void
     {
         $style = array(
             'cursor' => 'pointer',
@@ -59,14 +59,14 @@ class CreditsTest extends TestCase
 
         $this->credits->style($style);
         $this->assertEquals($style, $this->credits->style);
-        $this->assertRegExp('/"style":{"cursor":"pointer","color":"#909090","fontSize":"10px"}/', $this->chart->render());
+        $this->assertMatchesRegularExpression('/"style":{"cursor":"pointer","color":"#909090","fontSize":"10px"}/', $this->chart->render());
     }
 
-    public function testText()
+    public function testText(): void
     {
         $text = "Highcharts.com";
         $this->credits->text($text);
         $this->assertEquals($text, $this->credits->text);
-        $this->assertRegExp('/"text":"Highcharts.com"/', $this->chart->render());
+        $this->assertMatchesRegularExpression('/"text":"Highcharts.com"/', $this->chart->render());
     }
 }