+++ Tests/Highstock/RangeSelectorTest.php
@@ -7,7 +7,7 @@ use PHPUnit\Framework\TestCase;
 
 class RangeSelectorTest extends TestCase
 {
-    protected $chart;
+    protected Highstock $chart;
     protected $range;
 
     protected function setUp(): void
@@ -16,20 +16,20 @@ class RangeSelectorTest extends TestCase
         $this->range = $this->chart->rangeSelector;
     }
 
-    public function testButtonSpacing()
+    public function testButtonSpacing(): void
     {
         $spacing = 0;
         $this->range->buttonSpacing($spacing);
         $this->assertEquals($spacing, $this->range->buttonSpacing);
-        $this->assertRegExp('/"buttonSpacing":0/', $this->chart->render());
+        $this->assertMatchesRegularExpression('/"buttonSpacing":0/', $this->chart->render());
     }
 
-    public function testButtonTheme()
+    public function testButtonTheme(): void
     {
         $this->markTestIncomplete();
     }
 
-    public function testButtons()
+    public function testButtons(): void
     {
         $buttons = array(array(
             'type' => 'month',
@@ -39,101 +39,101 @@ class RangeSelectorTest extends TestCase
 
         $this->range->buttons($buttons);
         $this->assertEquals($buttons, $this->range->buttons);
-        $this->assertRegExp('/"buttons":\[{"type":"month","count":3,"text":"3m"}\]/', $this->chart->render());
+        $this->assertMatchesRegularExpression('/"buttons":\[{"type":"month","count":3,"text":"3m"}\]/', $this->chart->render());
     }
 
-    public function testEnabled()
+    public function testEnabled(): void
     {
         $this->range->enabled(true);
         $this->assertTrue($this->range->enabled);
-        $this->assertRegExp('/"enabled":true/', $this->chart->render());
+        $this->assertMatchesRegularExpression('/"enabled":true/', $this->chart->render());
 
         $this->range->enabled(false);
         $this->assertFalse($this->range->enabled);
-        $this->assertRegExp('/"enabled":false/', $this->chart->render());
+        $this->assertMatchesRegularExpression('/"enabled":false/', $this->chart->render());
     }
 
-    public function testInputBoxBorderColor()
+    public function testInputBoxBorderColor(): void
     {
         $color = 'silver';
         $this->range->inputBoxBorderColor($color);
         $this->assertEquals($color, $this->range->inputBoxBorderColor);
-        $this->assertRegExp('/"inputBoxBorderColor":"silver"/', $this->chart->render());
+        $this->assertMatchesRegularExpression('/"inputBoxBorderColor":"silver"/', $this->chart->render());
     }
 
-    public function testInputBoxHeight()
+    public function testInputBoxHeight(): void
     {
         $height = 16;
         $this->range->inputBoxHeight($height);
         $this->assertEquals($height, $this->range->inputBoxHeight);
-        $this->assertRegExp('/"inputBoxHeight":16/', $this->chart->render());
+        $this->assertMatchesRegularExpression('/"inputBoxHeight":16/', $this->chart->render());
     }
 
-    public function testInputBoxWidth()
+    public function testInputBoxWidth(): void
     {
         $width = 16;
         $this->range->inputBoxWidth($width);
         $this->assertEquals($width, $this->range->inputBoxWidth);
-        $this->assertRegExp('/"inputBoxWidth":16/', $this->chart->render());
+        $this->assertMatchesRegularExpression('/"inputBoxWidth":16/', $this->chart->render());
     }
 
-    public function testInputDateFormat()
+    public function testInputDateFormat(): void
     {
         $format = '%b %e, %Y';
         $this->range->inputDateFormat($format);
         $this->assertEquals($format, $this->range->inputDateFormat);
-        $this->assertRegExp('/"inputDateFormat":"%b %e, %Y"/', $this->chart->render());
+        $this->assertMatchesRegularExpression('/"inputDateFormat":"%b %e, %Y"/', $this->chart->render());
     }
 
-    public function testInputDateParser()
+    public function testInputDateParser(): void
     {
         $this->markTestIncomplete();
     }
 
-    public function testInputEditDateFormat()
+    public function testInputEditDateFormat(): void
     {
         $format = '%b %e, %Y';
         $this->range->inputEditDateFormat($format);
         $this->assertEquals($format, $this->range->inputEditDateFormat);
-        $this->assertRegExp('/"inputEditDateFormat":"%b %e, %Y"/', $this->chart->render());
+        $this->assertMatchesRegularExpression('/"inputEditDateFormat":"%b %e, %Y"/', $this->chart->render());
     }
 
-    public function testinputEnabled()
+    public function testinputEnabled(): void
     {
         $this->range->inputEnabled(true);
         $this->assertTrue($this->range->inputEnabled);
-        $this->assertRegExp('/"inputEnabled":true/', $this->chart->render());
+        $this->assertMatchesRegularExpression('/"inputEnabled":true/', $this->chart->render());
 
         $this->range->inputEnabled(false);
         $this->assertFalse($this->range->inputEnabled);
-        $this->assertRegExp('/"inputEnabled":false/', $this->chart->render());
+        $this->assertMatchesRegularExpression('/"inputEnabled":false/', $this->chart->render());
     }
 
-    public function testInputPosition()
+    public function testInputPosition(): void
     {
         $position= array(
             'align' => 'right'
         );
         $this->range->position($position);
         $this->assertEquals($position, $this->range->position);
-        $this->assertRegExp('/"position":{"align":"right"}/', $this->chart->render());
+        $this->assertMatchesRegularExpression('/"position":{"align":"right"}/', $this->chart->render());
     }
 
-    public function testInputStyle()
+    public function testInputStyle(): void
     {
         $this->markTestIncomplete();
     }
 
-    public function testLabelStyle()
+    public function testLabelStyle(): void
     {
         $this->markTestIncomplete();
     }
 
-    public function testSelected()
+    public function testSelected(): void
     {
         $index = 3;
         $this->range->selected($index);
         $this->assertEquals($index, $this->range->selected);
-        $this->assertRegExp('/"selected":3/', $this->chart->render());
+        $this->assertMatchesRegularExpression('/"selected":3/', $this->chart->render());
     }
 }